Refactor root home page
