+++ src/app/page.tsx
@@ -1,0 +1,5 @@
+import Homepage from "src/pages/Homepage/Homepage";
+
+export default function HomePage() {
+  return <Homepage />;
+}
